test aws region failover
