+++ tests/common_test_fixtures.py
@@ -1,11 +1,17 @@
 import base64
+import collections
 import os
 import pickle
 import tempfile
 import time
+import unittest
+import uuid
 
+import boto3
+import botocore.exceptions
 import fastapi
 from fastapi import testclient
+import moto
 import pandas as pd
 import pytest
 import requests
@@ -15,6 +21,9 @@ from sky import global_user_state
 from sky import sky_logging
 from sky.backends.cloud_vm_ray_backend import CloudVmRayBackend
 from sky.clouds.service_catalog import vsphere_catalog
+from sky.provision import common as provision_common
+from sky.provision.aws import config as aws_config
+from sky.provision.aws import instance as aws_instance
 from sky.provision.kubernetes import utils as kubernetes_utils
 from sky.serve import serve_state
 from sky.server import common as server_common
@@ -344,6 +353,109 @@ def mock_stream_utils(monkeypatch):
 
     monkeypatch.setattr("sky.server.stream_utils.stream_response",
                         _mock_stream_response)
+
+
+@pytest.fixture
+def mock_aws_backend(monkeypatch):
+    """Mock AWS backend for basic AWS testing operations."""
+    # Create a Subnet class to match what SkyPilot expects
+    Subnet = collections.namedtuple(
+        'Subnet',
+        ['subnet_id', 'vpc_id', 'availability_zone', 'map_public_ip_on_launch'])
+
+    # Mock subnet and VPC discovery
+    def mock_get_subnet_and_vpc_id(*args, **kwargs):
+        # Get the region from kwargs
+        region = kwargs.get('region', 'us-east-1')
+
+        # Create a subnet in the requested region
+        ec2 = boto3.resource('ec2', region_name=region)
+
+        # Create VPC and subnet in the requested region
+        vpc = ec2.create_vpc(CidrBlock='10.0.0.0/16')
+        subnet = ec2.create_subnet(VpcId=vpc.id,
+                                   CidrBlock='10.0.0.0/24',
+                                   AvailabilityZone=f"{region}a")
+
+        # Configure subnet to map public IPs
+        ec2.meta.client.modify_subnet_attribute(
+            SubnetId=subnet.id, MapPublicIpOnLaunch={'Value': True})
+
+        # Store subnet object
+        subnet_obj = Subnet(subnet_id=subnet.id,
+                            vpc_id=vpc.id,
+                            availability_zone=f"{region}a",
+                            map_public_ip_on_launch=True)
+
+        return ([subnet_obj], vpc.id)
+
+    # Mock security groups
+    def mock_get_or_create_vpc_security_group(*args, **kwargs):
+        # Return a mock security group
+        return unittest.mock.Mock(id="sg-12345678", group_name="test-sg")
+
+    # Mock IAM role
+    def mock_configure_iam_role(*args, **kwargs):
+        return {'Name': 'skypilot-test-role'}
+
+    def mock_wait_instances(region, cluster_name_on_cloud, state):
+        # Always return successfully without waiting
+        return
+
+    def mock_post_provision_runtime_setup(cloud_name, cluster_name,
+                                          cluster_yaml, provision_record,
+                                          custom_resource, log_dir):
+        # Get region from the provision record
+        region = provision_record.region
+
+        # Create a head instance
+        head_instance_id = f'i-{uuid.uuid4().hex[:8]}'
+
+        # Create instance info for the head
+        head_instance = provision_common.InstanceInfo(
+            instance_id=head_instance_id,
+            internal_ip='10.0.0.1',
+            external_ip='192.168.1.1',
+            tags={
+                'Name': cluster_name.name_on_cloud,
+                'ray-cluster-name': cluster_name.name_on_cloud,
+                'ray-node-type': 'head'
+            })
+
+        # Create ClusterInfo
+        instances = {head_instance_id: [head_instance]}
+        cluster_info = provision_common.ClusterInfo(
+            instances=instances,
+            head_instance_id=head_instance_id,
+            provider_name='aws',
+            provider_config={
+                'region': region,
+                'use_internal_ips': False
+            },
+            ssh_user='ubuntu')
+
+        return cluster_info
+
+    def mock_execute(self, handle, task, detach_run, dryrun=False):
+        # Return a fake job ID without attempting to SSH
+        return 1234
+
+    # Apply our mocks to the monkeypatch
+    monkeypatch.setattr(aws_config, '_get_subnet_and_vpc_id',
+                        mock_get_subnet_and_vpc_id)
+    monkeypatch.setattr(aws_config, '_get_or_create_vpc_security_group',
+                        mock_get_or_create_vpc_security_group)
+    monkeypatch.setattr(aws_config, '_configure_iam_role',
+                        mock_configure_iam_role)
+    monkeypatch.setattr(sky.provision.aws, 'wait_instances',
+                        mock_wait_instances)
+    # Add mock for post_provision_runtime_setup
+    monkeypatch.setattr(sky.provision.provisioner,
+                        'post_provision_runtime_setup',
+                        mock_post_provision_runtime_setup)
+    # Add mock for _execute
+    monkeypatch.setattr(sky.backends.cloud_vm_ray_backend.CloudVmRayBackend,
+                        '_execute', mock_execute)
 
 
 @pytest.fixture
