+++ tests/conftest.py
@@ -19,6 +19,7 @@ logger = sky_logging.init_logger(__name__)
 # tests can access them.
 from common_test_fixtures import aws_config_region
 from common_test_fixtures import enable_all_clouds
+from common_test_fixtures import mock_aws_backend
 from common_test_fixtures import mock_client_requests
 from common_test_fixtures import mock_controller_accessible
 from common_test_fixtures import mock_job_table_no_job
